Move code to models.py for Django 1.9 compatibility As of Django 1.9 it is no longer possible to run model imports in the root package
